refactor(geometry): split processors/swept.rs into disk + revolved
swept.rs (632 LOC) -> swept/{disk (SweptDiskSolid + RMF tube builder + tests,
355), revolved (RevolvedAreaSolid, 286), mod}. All under ~400. Pure code
motion; build_tube_rmf stays private; processors/mod.rs + siblings untouched.
build/clippy clean; swept/rmf/revolved tests green (27).
